NPl/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
tensorflow
opencv-python
nltk
spacy
textblob
emoji
cleantext
gensim
scipy==1.12
face-detection
ipython
transformers
huggingface_hub
gtts
langchain_community
pypdf
arxiv
pymupdf
google-generativeai
chromadb
langchain_google_genai
sentence-transformers
langchain_huggingface
tf-keras
faiss-cpu
langchain-chroma
langchain-text-splitters
langchain-core
newspaper3k
requests
lxml_html_clean
sqllit
